// apps/api/src/modules/identity/application/authorization/permission.enum.ts
export enum Permission {
  TENANT_READ = 'tenant.read',
  TENANT_WRITE = 'tenant.write',
  USER_READ = 'user.read',
  USER_WRITE = 'user.write',
  RESTAURANT_READ = 'restaurant.read',
  RESTAURANT_WRITE = 'restaurant.write',
  OUTLET_READ = 'outlet.read',
  OUTLET_WRITE = 'outlet.write',
  MENU_READ = 'menu.read',
  MENU_WRITE = 'menu.write',
  INVENTORY_READ = 'inventory.read',
  INVENTORY_WRITE = 'inventory.write',
  ORDER_READ = 'order.read',
  ORDER_WRITE = 'order.write',
  PAYMENT_READ = 'payment.read',
  PAYMENT_WRITE = 'payment.write',
  NOTIFICATION_READ = 'notification.read',
  NOTIFICATION_WRITE = 'notification.write',
  AUDIT_READ = 'audit.read',
  BILLING_READ = 'billing.read',
  BILLING_WRITE = 'billing.write',
}